# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.LoginUserInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.RegisterUserInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      role_id:
        type: integer
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - role_id
    - username
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API backend for digital parenting application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Digital Parenting App BE API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token upon successful login.
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful, returns JWT token
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Validation failed or invalid request body
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error during login
          schema:
            $ref: '#/definitions/models.Response'
      summary: User Login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: User Registration Details
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully, returns user ID
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "400":
          description: Validation failed or invalid request body
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "409":
          description: Username or Email already exists
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error during registration
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register New User
      tags:
      - Authentication
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer YOUR_JWT_TOKEN'' into the value field."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"